#2095 get event invoke config added
